{{cookiecutter.project_name}}/requirements.txt
alembic==1.4.3
Babel==2.8.0
Flask==1.1.2
Flask-Babel==2.0.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
flask_wtf
email-validator
Jinja2==2.11.2
flask-login
psycopg2-binary
Mako==1.1.3
MarkupSafe==1.1.1
python-dotenv==0.15.0
SQLAlchemy==1.3.20
Werkzeug
flask-marshmallow
marshmallow-sqlalchemy
flask-dance
sqlalchemy-utils
flask-restful
Flask-GraphQL
aniso8601==7.0.0
graphene==2.1.8
graphene-sqlalchemy>=2.0
flask-jwt-extended
factory-boy
